Renamed module API to Api
