+++ examples/x509auth_example.rb
@@ -13,8 +13,8 @@ USER_CERT_PASSWORD = 'mypassphrase'
 CA_PATH            = '/etc/grid-security/certificates'
 ENDPOINT           = 'https://localhost:3300'
 
-## get an OCCI::Client instance
-client = Occi::API::ClientHttp.new(ENDPOINT,
+## get an OCCI::Api::ClientHttp instance
+client = Occi::Api::ClientHttp.new(ENDPOINT,
         { :type               => "x509",
           :user_cert          => USER_CERT,
           :user_cert_password => USER_CERT_PASSWORD,
