+++ lib/occi/api/client/resource_output_factory.rb
@@ -2,7 +2,7 @@ require 'json'
 require 'erb'
 
 module Occi
-  module API
+  module Api
 
     class ResourceOutputFactory
