+++ spec/occi/api/client_amqp_spec.rb
@@ -1,172 +1,144 @@
-require 'rspec'
-require 'occi'
+require "rspec"
+require "occi"
+
+#Todo mock the amqp messages
 
 module Occi
-  module API
-
+  module Api
     describe ClientAmqp do
 
-      context "using media type text/plain" do
-        use_vcr_cassette "client_amqp_text_plain"
+      it "should do something" do
 
-        it "should establish connection" do
-          # TODO
-        end
-
-        it "should list all available resources" do
-        end
-
-        it "should list compute resources" do
-        end
-
-        it "should list network resources" do
-        end
-
-        it "should list storage resources" do
-        end
-
-        it "should list all available mixins" do
-        end
-
-        it "should list os_tpl mixins" do
-        end
-
-        it "should list resource_tpl mixins" do
-        end
-
-        it "should describe all available resources" do
-        end
-
-        it "should describe compute resources" do
-        end
-
-        it "should describe network resources" do
-        end
-
-        it "should describe storage resources" do
-        end
-
-        it "should describe all available mixins" do
-        end
-
-        it "should describe os_tpl mixins" do
-        end
-
-        it "should describe resource_tpl mixins" do
-        end
-
-        it "should create a new compute resource" do
-        end
-
-        it "should create a new storage resource" do
-        end
-
-        it "should create a new network resource" do
-        end
-
-        it "should delete a compute resource" do
-        end
-
-        it "should delete a network resource" do
-        end
-
-        it "should delete a storage resource" do
-        end
-
-        it "should trigger an action on a compute resource" do
-        end
-
-        it "should trigger an action on a storage resource" do
-        end
-
-        it "should trigger an action on a network resource" do
-        end
-
-        it "should refresh its model" do
-        end
+        # TODO: Implement scenarios for client
+      end
+=begin
+      before(:all) do
+        @client = Occi::Api::ClientAmqp.new("http://localhost:9292/", auth_options = { :type => "none" },
+                                            log_options = { :out => STDERR, :level => Occi::Log::WARN, :logger => nil },
+                                            media_type = "application/occi+json")
       end
 
-      context "using media type application/occi+json" do
-        use_vcr_cassette "client_amqp_application_occi_json"
+      it "initialize and connect client" do
+        @client.connected.should be_true
 
-        it "should establish connection" do
-          # TODO
-        end
+        @client.model.actions  .should have_at_least(1).actions
+        @client.model.kinds    .should have_at_least(3).kinds
+        @client.model.mixins   .should have_at_least(1).mixins
+        @client.model.links    .should be_empty
+        @client.model.resources.should be_empty
 
-        it "should list all available resources" do
-        end
-
-        it "should list compute resources" do
-        end
-
-        it "should list network resources" do
-        end
-
-        it "should list storage resources" do
-        end
-
-        it "should list all available mixins" do
-        end
-
-        it "should list os_tpl mixins" do
-        end
-
-        it "should list resource_tpl mixins" do
-        end
-
-        it "should describe all available resources" do
-        end
-
-        it "should describe compute resources" do
-        end
-
-        it "should describe network resources" do
-        end
-
-        it "should describe storage resources" do
-        end
-
-        it "should describe all available mixins" do
-        end
-
-        it "should describe os_tpl mixins" do
-        end
-
-        it "should describe resource_tpl mixins" do
-        end
-
-        it "should create a new compute resource" do
-        end
-
-        it "should create a new storage resource" do
-        end
-
-        it "should create a new network resource" do
-        end
-
-        it "should delete a compute resource" do
-        end
-
-        it "should delete a network resource" do
-        end
-
-        it "should delete a storage resource" do
-        end
-
-        it "should trigger an action on a compute resource" do
-        end
-
-        it "should trigger an action on a storage resource" do
-        end
-
-        it "should trigger an action on a network resource" do
-        end
-
-        it "should refresh its model" do
-        end
       end
 
+      it "create compute" do
+        res = Occi::Infrastructure::Compute.new
+        res.title = "MyComputeResource1"
+        res.mixins << @client.find_mixin('small', "resource_tpl")
+        res.mixins << @client.find_mixin('my_os', "os_tpl")
+        #
+        uri_new = @client.create res
+        uri_new.should include('/compute/')
+        @client.last_response_status.should == 201
+
+        res.title = "MyComputeResource2"
+        uri_new = @client.create res
+        uri_new.should include('/compute/')
+        @client.last_response_status.should == 201
+
+      end
+
+      it "list /" do
+        list = @client.list
+        list.should have_at_least(1).list
+      end
+
+      it "list compute" do
+        list = @client.list "compute"
+        list.should have_at_least(1).list
+      end
+
+      it "describe /" do
+        description = @client.describe
+
+        description[0].resources.should have_at_least(1).resources
+      end
+
+      it "describe compute" do
+        description = @client.describe "compute"
+
+        description[0].resources.should have_at_least(1).resources
+      end
+
+      it "describe compute/uuid" do
+        list = @client.list "compute"
+        list.should have_at_least(1).list
+
+        compute = list.first
+
+        description = @client.describe compute
+
+        description[0].resources.should have_at_least(1).resources
+        resource = description[0].resources.first
+
+        resource_type_identifier = @client.endpoint.chomp('/') + resource.kind.location + resource.id
+        resource_type_identifier.should == compute.to_s
+      end
+
+      it "trigger compute stop" do
+        list = @client.list "compute"
+        list.should have_at_least(1).list
+
+        list.each do |key, value|
+          description = @client.describe key
+          @compute = key
+          @resource    = description[0].resources.first
+
+          if @resource.attributes.occi.compute.state == "active"
+            break
+          end
+        end
+
+        @resource.actions.each do |key , value|
+          if key.term == "stop"
+            @action = key
+            break
+          end
+        end
+
+        @action.should_not be_nil, "action is nil"
+        @action.term.should == "stop"
+
+        @client.trigger(@compute, @action.to_s)
+
+        description = @client.describe @compute
+        @resource   = description[0].resources.first
+
+        @resource.attributes.occi.compute.state.should == "inactive"
+      end
+
+      it "delete compute/uuid" do
+        list = @client.list "compute"
+        list_count = list.count
+        list_count.should >= 1
+
+        compute = list.first
+
+        response = @client.delete compute
+        response.should == true
+        @client.last_response_status.should == 200
+
+       list = @client.list "compute"
+        list.count.should < list_count
+      end
+
+      it "delete /" do
+        response = @client.delete "/"
+        response.should == true
+        @client.last_response_status.should == 200
+
+        list = @client.list "compute"
+        list.count.should == 0
+      end
+=end
     end
-
   end
 end
