+++ spec/spec_helper.rb
@@ -15,22 +15,24 @@ RSpec.configure do |c|
 end
 
 module Occi
-  module API
+  module Api
+    module Helpers
 
-    def self.conn_helper
-      #Occi::API::ClientHttp.new('https://localhost:3300',
-      #                          { :type  => "none" },
-      #                          { :out   => "/dev/null",
-      #                            :level => Occi::Log::DEBUG })
+      def self.conn_helper
+        #Occi::Api::ClientHttp.new('https://localhost:3300',
+        #                          { :type  => "none" },
+        #                          { :out   => "/dev/null",
+        #                            :level => Occi::Log::DEBUG })
 
-      Occi::API::ClientHttp.new("https://carach5.ics.muni.cz:11443",
-        { :type               => "x509",
-          :user_cert          => ENV['HOME'] + '/.globus/usercred.pem',
-          :user_cert_password => "",
-          :ca_path            => '/etc/grid-security/certificates' },
-        { :out   => STDERR,
-          :level => Occi::Log::DEBUG })
+        Occi::Api::ClientHttp.new("https://carach5.ics.muni.cz:11443",
+          { :type               => "x509",
+            :user_cert          => ENV['HOME'] + '/.globus/usercred.pem',
+            :user_cert_password => "",
+            :ca_path            => '/etc/grid-security/certificates' },
+          { :out   => STDERR,
+            :level => Occi::Log::DEBUG })
+      end
+
     end
-
   end
 end
